refact : qna/edit 관련 구조 변경
