+++ client/src/features/question-composer/edit/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as QuestionEditForm } from './QuestionEditForm';
